// packages/backend/src/usagers/CerfaFields.type.ts
// Liste des champs disponibles dans les Cerfa de demande et d'attestation
export type CerfaFields = {
  adresse: string;
  adresseOrga1: string;
  agrement: string;
  anneeDebut: string;
  anneeDecision1A: string;
  anneeDecision1B: string;
  anneeDecision2: string;
  anneeFin: string;
  anneeNaissance1: string;
  anneeNaissance2: string;
  anneePremiereDom: string;
  anneeRdv: string;
  ayantsDroits: string;
  courriel: string;
  courrielOrga: string;
  decision: string;
  entretienAdresse: string;
  entretienAvec: string;
  heureRdv: string;
  jourDebut: string;
  jourDecision1A: string;
  jourDecision1B: string;
  jourDecision2: string;
  jourFin: string;
  jourNaissance1: string;
  jourNaissance2: string;
  jourPremiereDom: string;
  jourRdv: string;
  lieuNaissance1: string;
  lieuNaissance2: string;
  minutesRdv: string;
  moisDebut: string;
  moisDecision1A: string;
  moisDecision1B: string;
  moisDecision2: string;
  moisFin: string;
  moisNaissance1: string;
  moisNaissance2: string;
  moisPremiereDom: string;
  moisRdv: string;
  motifRefus: string;
  nomOrga1: string;
  nomOrga2: string;
  noms1: string;
  noms2: string;
  numeroUsager: string;
  orientation: string;
  prefecture1: string;
  prefecture2: string;
  prenoms1: string;
  prenoms2: string;
  rattachement: string;
  responsable: string;
  sexe1: string;
  sexe2: string;
  signature1A: string;
  signature1B: string;
  signature2: string;
  telephone: string;
  telephoneOrga: string;
  typeDemande: string;
};